+++ docs/overview/objects.php
@@ -1,6 +1,6 @@
 <?php /*~*/
 use NitroXy\PHPForms\Form;
-use NitroXy\PHPForms\FormSelect;
+use NitroXy\PHPForms\FormOptions;
 ?>
 <?php
 
@@ -17,6 +17,6 @@ $myObject->description = 'Lorem ipsum dot sit amet.';
 Form::from_object($myObject, function($f){
 	$f->text_field('name', 'Name');
 	$f->text_field('age', 'Age', ['type' => 'number', 'min' => 1]);
-	$f->select(FormSelect::from_array($f, 'role', array('', 'Manager', 'Frobnicator', 'Developer'), 'Role'));
+	$f->select('role', 'Role', FormOptions::from_array(['', 'Manager', 'Frobnicator', 'Developer']));
 	$f->textarea('description', 'Description');
 }, ['layout' => 'bootstrap']);