+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Breaking changes
 
+  - Select API change: prototype more similar to other fields, replaced
+    `FormSelect` with `FormOptions`.
   - Button API change: $key removed from prototype.
 
 # 1.4.0